removed logic from api/schema
