+++ types/index.ts
@@ -1,0 +1,15 @@
+export interface DataAttribute<T> {
+  data: {
+    attributes: T;
+  };
+}
+
+export interface DataAttributes<T> {
+  data: {
+    attributes: T;
+  }[];
+}
+
+export interface CategoryResponse {
+  type: string;
+}
